[refactor/move_example2_activity] Moving the Example2Activity into the example1 module. I'm going to keep all the activities related to the same kind of Rx operation on the same package.
