+++ core/src/main/java/com/kimboo/core/activities/ActivityBuilder.kt
@@ -6,10 +6,10 @@ import android.content.Intent
 class ActivityBuilder {
     companion object {
         fun getExample1Intent(context: Context) : Intent {
-            return Intent(context, Class.forName("com.kimboo.example1.ui.Example1Activity"))
+            return Intent(context, Class.forName("com.kimboo.example1.ui.examplezip1.Example1Activity"))
         }
         fun getExample2Intent(context: Context) : Intent {
-            return Intent(context, Class.forName("com.kimboo.example2.ui.Example2Activity"))
+            return Intent(context, Class.forName("com.kimboo.example1.ui.examplezip2.Example2Activity"))
         }
     }
 }
