+++ example1/src/main/AndroidManifest.xml
@@ -2,7 +2,9 @@
     package="com.kimboo.example1">
 
     <application>
-        <activity android:name=".ui.Example1Activity"
+        <activity android:name=".ui.examplezip1.Example1Activity"
+            android:launchMode="singleTop" />
+        <activity android:name=".ui.examplezip2.Example2Activity"
             android:launchMode="singleTop" />
     </application>
