+++ example1/src/main/java/com/kimboo/example1/di/component/Example1ViewInjector.kt
@@ -2,7 +2,8 @@ package com.kimboo.example1.di.component
 
 import com.kimboo.di.component.BaseSubComponent
 import com.kimboo.example1.di.modules.Example1ViewModelModule
-import com.kimboo.example1.ui.Example1Activity
+import com.kimboo.example1.ui.examplezip1.Example1Activity
+import com.kimboo.example1.ui.examplezip2.Example2Activity
 import dagger.Component
 import javax.inject.Singleton
 
@@ -11,7 +12,10 @@ import javax.inject.Singleton
     modules = [Example1ViewModelModule::class]
 )
 interface Example1ViewInjector {
+
     fun inject(activity: Example1Activity)
+    fun inject(activity: Example2Activity)
+
     @Component.Builder
     interface Builder {
         fun baseSubComponent(baseSubComponent: BaseSubComponent): Builder
