+++ example1/src/main/java/com/kimboo/example1/di/modules/Example1ViewModelModule.kt
@@ -1,7 +1,8 @@
 package com.kimboo.example1.di.modules
 
 import androidx.lifecycle.ViewModel
-import com.kimboo.example1.ui.viewmodel.Example1ViewModel
+import com.kimboo.example1.ui.examplezip1.viewmodel.Example1ViewModel
+import com.kimboo.example1.ui.examplezip2.viewmodel.Example2ViewModel
 import com.kimboo.utils.ViewModelKey
 import dagger.Binds
 import dagger.Module
@@ -14,5 +15,12 @@ abstract class Example1ViewModelModule {
     @ViewModelKey(Example1ViewModel::class)
     abstract fun bindExample1ViewModel(
         example1ViewModel: Example1ViewModel
+    ): ViewModel
+
+    @Binds
+    @IntoMap
+    @ViewModelKey(Example2ViewModel::class)
+    abstract fun bindExample2ViewModel(
+        example2ViewModel: Example2ViewModel
     ): ViewModel
 }
