+++ example1/src/main/res/layout/activity_example_2.xml
