+++ example2/src/main/java/com/kimboo/example2/di/component/Example2ViewInjector.kt
@@ -1,7 +1,6 @@
 package com.kimboo.example2.di.component
 
 import com.kimboo.di.component.BaseSubComponent
-import com.kimboo.example2.ui.Example2Activity
 import com.kimboo.example2.di.modules.Example2ViewModelModule
 import dagger.Component
 import javax.inject.Singleton
@@ -11,7 +10,6 @@ import javax.inject.Singleton
     modules = [Example2ViewModelModule::class]
 )
 interface Example2ViewInjector {
-    fun inject(activity: Example2Activity)
     @Component.Builder
     interface Builder {
         fun baseSubComponent(baseSubComponent: BaseSubComponent): Builder
